% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMLExpData}
\alias{AMLExpData}
\title{AMLExpData}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
AMLExpData
}
\description{
AML exp data
}
\keyword{datasets}
