% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CSCE_ProcExp}
\alias{CSCE_ProcExp}
\title{CSCE_ProcExp}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
CSCE_ProcExp
}
\description{
CSCE_ProcExp
}
\keyword{datasets}
