% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CSCEExpData}
\alias{CSCEExpData}
\title{CSCEExpData}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
CSCEExpData
}
\description{
CSCE exp data
}
\keyword{datasets}
