% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DNARepairGene}
\alias{DNARepairGene}
\title{DNARepairGene}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
DNARepairGene
}
\description{
DNARepairGene
}
\keyword{datasets}
