% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DNARepairGMTDataList}
\alias{DNARepairGMTDataList}
\title{DNA repair gene}
\format{
A vector
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
DNARepairGMTDataList
}
\description{
A gene list of DNA repair
}
\keyword{datasets}
