% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EmptyDF}
\alias{EmptyDF}
\title{EmptyDF}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
EmptyDF
}
\description{
EmptyDF
}
\keyword{datasets}
