% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRGEXCOR.R
\name{Func_CorPlot_v2}
\alias{Func_CorPlot_v2}
\title{Title Func_CorPlot_v2}
\usage{
Func_CorPlot_v2(
  ExpData = CSCE_ProcExp[[DNARepairGene]],
  YAxisGene = DNARepairGene,
  XAxisGene = "TREM1",
  CancerType = "CSCE"
)
}
\arguments{
\item{ExpData}{Gene exp data frame.}

\item{YAxisGene}{DNA repair}

\item{XAxisGene}{TREM1}

\item{CancerType}{}
}
\value{
a plot
}
\description{
Title Func_CorPlot_v2
}
\examples{
Func_CorPlot_v2(ExpData=CSCE_ProcExp[[DNARepairGene]],  YAxisGene=DNARepairGene, XAxisGene="TREM1", CancerType="CSCE" )
}
