% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRGEXCOR.R
\name{Func_CorrPlotInputDF}
\alias{Func_CorrPlotInputDF}
\title{Title To make data.frame to be input for plot}
\usage{
Func_CorrPlotInputDF(
  DNARepairGene = "AAAS",
  XAxisGene = "TREM1",
  MyExpData = CSCEExpData
)
}
\arguments{
\item{DNARepairGene}{My interest of DNA repair gene}

\item{XAxisGene}{TREM1}

\item{MyExpData}{Expression data.frame}
}
\value{
Processed data.frame to be input for plot
}
\description{
Title To make data.frame to be input for plot
}
\examples{
Func_CorrPlotInputDF(DNARepairGene="AAAS",XAxisGene="TREM1", MyExpData=CSCEExpData)
}
