% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRGEXCOR.R
\name{Func_EmptyPlot}
\alias{Func_EmptyPlot}
\title{Title Func_EmptyPlot}
\usage{
Func_EmptyPlot(EmptyDF, Xaxis, Yaxis, YAxisGene = DNARepairGene)
}
\arguments{
\item{EmptyDF}{When no DNA repaire gene}

\item{Xaxis}{TREM1}

\item{Yaxis}{DNA repair gene}

\item{YAxisGene}{DNARepairGene}
}
\value{
an empty plot
}
\description{
Title Func_EmptyPlot
}
\examples{
Func_EmptyPlot(EmptyDF, Xaxis,Yaxis, YAxisGene=DNARepairGene)
}
