% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LHCExpData}
\alias{LHCExpData}
\title{LHCExpData}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
LHCExpData
}
\description{
LHC exp data
}
\keyword{datasets}
