% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSCC_ProcExp}
\alias{LSCC_ProcExp}
\title{LSCC_ProcExp}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
LSCC_ProcExp
}
\description{
LSCC_ProcExp
}
\keyword{datasets}
