% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSCCExpData}
\alias{LSCCExpData}
\title{LSCCExpData}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
LSCCExpData
}
\description{
LSCC exp data
}
\keyword{datasets}
