% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MyExpData_RowHugoSymb}
\alias{MyExpData_RowHugoSymb}
\title{MyExpData_RowHugoSymb}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
MyExpData_RowHugoSymb
}
\description{
MyExpData_RowHugoSymb,
}
\keyword{datasets}
