% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NumberVector}
\alias{NumberVector}
\title{NumberVector}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
NumberVector
}
\description{
NumberVector
}
\keyword{datasets}
