% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRGEXCOR.R
\name{ShinySever}
\alias{ShinySever}
\title{Title ShinySever}
\usage{
ShinySever(input, output)
}
\arguments{
\item{input}{DNA repaire gene}

\item{output}{a plot}
}
\value{
shiny app
}
\description{
Title ShinySever
}
\examples{
ShinySever(input, output)
}
