% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UserInterface}
\alias{UserInterface}
\title{UserInterface}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
UserInterface
}
\description{
UserInterface
}
\keyword{datasets}
