% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XAxisGene}
\alias{XAxisGene}
\title{XAxisGene}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/sanghoonleeeins/DRGEXCOR}
}
\usage{
XAxisGene
}
\description{
XAxisGene, TREM1
}
\keyword{datasets}
