% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRGEXCOR.R
\name{cal_z_score}
\alias{cal_z_score}
\title{z-score calibratin}
\usage{
cal_z_score(MyExpData_RowHugoSymb)
}
\arguments{
\item{x}{data.frame input}
}
\value{
z-score data.frame
}
\description{
z-score calibratin
}
\examples{
cal_z_score(MyExpData_RowHugoSymb)
}
